requirements.txt
py-clob-client>=0.17.0
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.7.0
anthropic>=0.40.0
